Summarizing Functional Linear Model Fits

summary method for class "IWTlm". Function returning a summary of the results of IWT for the test on a functional linear model: minimum IWT-adjusted p-values of the F-tests on the whole model and of t-tests on all covariates' effects are reported.

Usage

# S3 method for class 'IWTlm'
summary(object, ...)

# S3 method for class 'TWTlm'
summary(object, ...)

Arguments

object

An object of class "IWTlm", usually, a result of a call to IWTlm.

...

Further arguments passed to or from other methods.

Value

No value returned. The function summary.IWTlm computes and returns a list of summary statistics of the fitted functional analysis of variance given in object, using the component "call" from its arguments, plus:

ttest: A L+1 x 1 matrix with columns for the functional regression coefficients, and corresponding (two-sided) IWT-adjusted minimum p-values of t-tests (i.e., the minimum p-value over all p basis components used to describe functional data).

R2: Range of the functional R-squared.

ftest: IWT-adjusted minimum p-value of functional F-test.

References

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data. Journal of Nonparametric Statistics, 29(2), 407-424.

Pini, A., Vantini, S., Colosimo, B. M., & Grasso, M. (2018). Domain‐selective functional analysis of variance for supervised statistical profile monitoring of signal data. Journal of the Royal Statistical Society: Series C (Applied Statistics) 67(1), 55-81.

Abramowicz, K., Hager, C. K., Pini, A., Schelin, L., Sjostedt de Luna, S., & Vantini, S. (2018). Nonparametric inference for functional‐on‐scalar linear models applied to knee kinematic hop data after injury of the anterior cruciate ligament. Scandinavian Journal of Statistics 45(4), 1036-1061.

See also

Examples

# Importing the NASA temperatures data set
data(NASAtemp)

temperature <- rbind(NASAtemp$milan, NASAtemp$paris)
groups <- c(rep(0, 22), rep(1, 22))

# Performing the IWT
IWT.result <- IWTlm(temperature ~ groups, B = 2L)
#> Error in eval(predvars, data, env): object 'groups' not found

# Summary of the IWT results
summary(IWT.result)
#> Error: object 'IWT.result' not found

# Plot of the IWT results
layout(1)
plot(IWT.result)
#> Error: object 'IWT.result' not found

# All graphics on the same device
layout(matrix(1:4, nrow = 2, byrow = FALSE))
plot(
  IWT.result, 
  main = 'NASA data', 
  plot_adjpval = TRUE, 
  xlab = 'Day', 
  xrange = c(1, 365)
)
#> Error: object 'IWT.result' not found